05272002
修改admin学生管理页面删除的功能，学生有导师后将不能随意删除，课题管理删除 删除键

+++ WebServer--graduate/myhome/views/AdminViews.py
@@ -108,8 +108,11 @@ def student_list(request):
 
 def doDeleteStudent(request):
     data = request.POST.dict()
-    Student.objects.filter(id = data['id']).delete()
-    return JsonResponse({'msg':"删除成功",'code':200})
+    if StudentSelectTitle.objects.filter(student_id=data['id']).exists():
+        return JsonResponse({'msg': "无法删除,已有指导老师'", 'code': 201})
+    else:
+        Student.objects.filter(id = data['id']).delete()
+        return JsonResponse({'msg':"删除成功",'code':200})
 
 
 def doSearchStudent(request):
